refactor: organize heritageCalculations into themed calculation modules (Phase 4 - part 2)
**Created:**
- src/utils/calculations/glowContributions.ts (136 lines)
  - calculateGlowContribution - Heritage glow value calculation
  - getGlowReductionPercentage - Damage-based glow reduction
  - getAgeColorCode - Age-based marker colors

- src/utils/calculations/heritageMetrics.ts (73 lines)
  - calculateTotalHeritageValue - Total heritage value sum
  - calculateDestroyedValue - Destroyed value by date
  - calculateHeritageIntegrity - Integrity percentage calculation

- src/utils/calculations/significance.ts (30 lines)
  - calculateSignificanceScore - Marker size calculation

- src/utils/calculations/index.ts (17 lines)
  - Barrel export for themed modules

**Modified:**
- src/utils/heritageCalculations.ts (219 → 19 lines, 91% reduction)
  - Now re-exports from calculations/ for backward compatibility
  - Marked as @deprecated with migration guidance
  - All existing imports continue to work

**Benefits:**
- Clear thematic organization (glow, metrics, significance)
- Each module focused on specific calculation type
- Easy to find and modify specific calculations
- Backward compatible (no breaking changes)
- Better for future timeline animation development

**Testing:**
- All 184 tests passing
- Linter clean (0 errors, 0 warnings)
- All existing imports work via barrel export
